important bugfix for hint prompter
