+++ src/tmtccmd/utility/json_handler.py
@@ -49,3 +49,18 @@ def check_json_file(json_cfg_path: str) -> bool:
                 json.dump(void_data, file)
                 return False
     return True
+
+
+def save_to_json_with_prompt(
+        key: str, value: any, name: str, json_cfg_path: str, json_obj: any
+) -> bool:
+    logger = get_console_logger()
+    save_to_json = input(
+        f'Do you want to store the {name} to the configuration file? (y/n): '
+    )
+    if save_to_json.lower() in ['y', 'yes']:
+        json_obj[key] = value
+        logger.info(f'The {name} was stored to the JSON file {json_cfg_path}')
+        logger.info('Delete this file or edit it manually to change it')
+        return True
+    return False
